+++ src/lib/i18n/resources.ts
@@ -1,24 +1,32 @@
 import adminEn from "./resources/en/admin.json";
+import botEn from "./resources/en/bot.json";
 import commonEn from "./resources/en/common.json";
 import dashboardEn from "./resources/en/dashboard.json";
+import extensionEn from "./resources/en/extension.json";
 import homeEn from "./resources/en/home.json";
 import playlistEn from "./resources/en/playlist.json";
 import searchEn from "./resources/en/search.json";
 import adminEs from "./resources/es/admin.json";
+import botEs from "./resources/es/bot.json";
 import commonEs from "./resources/es/common.json";
 import dashboardEs from "./resources/es/dashboard.json";
+import extensionEs from "./resources/es/extension.json";
 import homeEs from "./resources/es/home.json";
 import playlistEs from "./resources/es/playlist.json";
 import searchEs from "./resources/es/search.json";
 import adminFr from "./resources/fr/admin.json";
+import botFr from "./resources/fr/bot.json";
 import commonFr from "./resources/fr/common.json";
 import dashboardFr from "./resources/fr/dashboard.json";
+import extensionFr from "./resources/fr/extension.json";
 import homeFr from "./resources/fr/home.json";
 import playlistFr from "./resources/fr/playlist.json";
 import searchFr from "./resources/fr/search.json";
 import adminPtBr from "./resources/pt-br/admin.json";
+import botPtBr from "./resources/pt-br/bot.json";
 import commonPtBr from "./resources/pt-br/common.json";
 import dashboardPtBr from "./resources/pt-br/dashboard.json";
+import extensionPtBr from "./resources/pt-br/extension.json";
 import homePtBr from "./resources/pt-br/home.json";
 import playlistPtBr from "./resources/pt-br/playlist.json";
 import searchPtBr from "./resources/pt-br/search.json";
@@ -26,32 +34,40 @@ import searchPtBr from "./resources/pt-br/search.json";
 export const i18nResources = {
   en: {
     admin: adminEn,
+    bot: botEn,
     common: commonEn,
     dashboard: dashboardEn,
+    extension: extensionEn,
     home: homeEn,
     playlist: playlistEn,
     search: searchEn,
   },
   es: {
     admin: adminEs,
+    bot: botEs,
     common: commonEs,
     dashboard: dashboardEs,
+    extension: extensionEs,
     home: homeEs,
     playlist: playlistEs,
     search: searchEs,
   },
   fr: {
     admin: adminFr,
+    bot: botFr,
     common: commonFr,
     dashboard: dashboardFr,
+    extension: extensionFr,
     home: homeFr,
     playlist: playlistFr,
     search: searchFr,
   },
   "pt-BR": {
     admin: adminPtBr,
+    bot: botPtBr,
     common: commonPtBr,
     dashboard: dashboardPtBr,
+    extension: extensionPtBr,
     home: homePtBr,
     playlist: playlistPtBr,
     search: searchPtBr,